requirements.txt
# UI and certificate workflow dependencies
streamlit>=1.35
python-dotenv>=1.0
reportlab>=3.6.0
pillow>=9.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
pandas>=1.3.0
openpyxl>=3.0.0
pypdf>=4,<5
